refactor: improve code and path structure
